requirements.txt
openai
python-dotenv
requests
beautifulsoup4
pandas
numpy
farm-haystack[inmemory]
